oc: make route utility reusable across packages
